deleting room now transfers equipment from that room to main room

+++ Projekat_SIMS_IN_TIM3/Projekat_SIMS_IN_TIM3/ManagerWindows/RoomWindow.xaml.cs
@@ -56,7 +56,16 @@ namespace Projekat_SIMS_IN_TIM3.ManagerWindows
         private void Delete_Click(object sender, RoutedEventArgs e)
         {
             Room room = (Room)((Button)e.Source).DataContext;
+            if(room.Id == 0)
+            {
+                MessageBox.Show("Default storage room cant be deleted!");
+                return;
+            }
             Rooms.Remove(room);
+            foreach (var equipment in EquipmentWindow.Equipment_All.Where(x => x.RoomId == room.Id))
+            {
+                equipment.RoomName= "Main Storage";
+            }
             this.roomController.DeleteById(room.Id);
         }
